preprocessing/frames_generator/requirements.txt
facenet-pytorch==2.5.3
numpy==1.26.4
opencv-python==4.8.1.78
pandas==2.2.1
pillow==10.2.0
torch==2.2.2
tqdm==4.66.2
